M2-5204: Move all session applets fixtures in one place
