+++ src/apps/activities/tests/fixtures/activities.py
@@ -11,8 +11,3 @@ def activity_create(single_select_item_create: ActivityItemCreate) -> ActivityCr
     return ActivityCreate(
         name="test", description={Language.ENGLISH: "test"}, items=[single_select_item_create], key=uuid.uuid4()
     )
-
-
-@pytest.fixture(scope="session")
-def activity_create_session(item_create: ActivityItemCreate) -> ActivityCreate:
-    return ActivityCreate(name="test", description={Language.ENGLISH: "test"}, items=[item_create], key=uuid.uuid4())
